backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic[email]
python-multipart==0.0.6
python-dotenv==1.0.1
twilio==9.0.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.12.0
lightgbm==4.3.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
